+++ NEWS.md
@@ -1,3 +1,13 @@
+# Transition 1.0.1
+
+* Improve `adjust()` formula (#9).
+
+* Allow `result` arg of `add_transitions()` and `get_transitions()` to refer to an integer (or numeric) vector of binary data with values either 1 or 0 (#10).
+
+* Add parameter `prev_result` to `add_ prev_result()` (#11, #12)
+
+* Add parameter `prev_date` to `add_ prev_date()` (#13, #14)
+
 # Transition 1.0.0
 
 * Initial CRAN submission.